ml/requirements.txt
# ─── Core LLM ────────────────────────────────────────
llama-cpp-python>=0.2.90

# ─── API Server ──────────────────────────────────────
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0

# ─── Speech-to-Text ──────────────────────────────────
faster-whisper>=1.0.0

# ─── Text-to-Speech ──────────────────────────────────
piper-tts>=1.2.0
pyttsx3>=2.90

# ─── ML / Deep Learning ─────────────────────────────
torch>=2.1.0
torchaudio>=2.1.0
transformers>=4.30.0
sentence-transformers>=2.2.0
accelerate>=0.20.0
safetensors>=0.4.0

# ─── NLP ─────────────────────────────────────────────
spacy>=3.7.0

# ─── Audio Processing ───────────────────────────────
librosa>=0.10.0
sounddevice>=0.4.6
soundfile>=0.12.0

# ─── Computer Vision (Proctoring) ───────────────────
opencv-python>=4.8.0

# ─── RAG / Vector DB ────────────────────────────────
chromadb>=0.4.0

# ─── ATS / Resume Parsing ───────────────────────────
scikit-learn>=1.4.0
pypdf>=4.0.0

# ─── Utilities ───────────────────────────────────────
python-dotenv>=1.0.0
requests>=2.31.0
tabulate>=0.9.0
rich>=13.0.0
coloredlogs>=15.0

# ─── Data Engineering / Dataset Processing ──────────
kagglehub>=0.2.0
pandas>=2.0.0
datasets>=2.14.0
